Fixed function calls in expression parser with >1 argument
Not sure how/if this ever worked in the past...commas were not being
added to the parameter stack, which resulted in only the final argument
being passed into EvaluateFunction(). The good news is that since this
code is obviously not being exercised (if it was, we would have realised
sooner), this change *shouldn't* break any existing files.

Note that I'm assuming that EvaluateFunction()'s arguments are supposed
to be in prefix order. Since all supported functions only accept a
single argument, this won't actually have an effect currently. Note that
I've changed the divide function to accept more than one argument now,
so going forward, it will assume prefix order.