Fixed some graphs in MicroClimate validation file
The graphing code will throw if an integer column contains blank values.
This occurs in this graph because it's in a multi-zone simulation and
some of the report outputs are specific to a particular zone. The
solution is to add a filter to the affected graph series so that they
exclude results from the other zones.