Fixed UI file closing logic
Previously, presenters were not being removed from the Presenters1 and
presenters2 lists after being detached when clicking the close button
for the entire GUI. This causes a number of problems if the GUI is still
open after this (which can happen if multiple files were open but the
cancel button is clicked when prompted to save one of them). The most
obvious problem is that after this happens, clicking the close button
will cause a null reference exception to be thrown, preventing the GUI
from closing.