Disabled clem summary outputs to get jenkins builds green again
The problem appears to be writing to the same file from multiple threads
concurrently. This occurs when running multiple clem files in parallel -
or possibly running multiple simulations in the same file in parallel.
The problem doesn't occur every time of course - it only happens if >1
simulation gets to this point at the same time. I've temporarily
disabled this feature to get jenkins builds green again.