Improvements to the gtk3 editorview
This view needs to be able to be initialised via the ViewBase glade file
infrastructure via the overriden Initialise() method.

I've also moved the roslyn intellisense initialisation out of the
constructor and into the EditorType setter, so that we only perform this
initialisation if the editor is being used for a manager script.