Added an option to manager UI to change style schemes
This is necessary because the gtksourceview colour scheme does not
necessarily work well with the system-wide gtk theme. E.g. if using a
dark gtk theme, the defualt stylescheme for the gtksourceview
('classic') uses white as the current line highlighter colour.

Any changes to the style scheme here will persist between restarts of
apsim (I'm saving the style scheme's ID to the user settings).