Fixed space allocation in series UI
Previously, both the properties grid and the plot view were both packed
with expand = TRUE, which was causing the properties grid to be
allocated any extra space (which is just wasted as there's nothing to
show in the empty space). We now set expand = FALSE on the properties
grid, so that any extra space goes to the plotview.