Fixed the appearance of series UI
The SeriesView is constructed from a glade file. The gtk2 version uses a
now-obsolete GtkTable widget. The way that I've setup the gtk3 version
of seriesview is a bit weird, we use the same glade file but remove the
table and replace it with a GtkGrid, and manually pack in all of the
labels which were present in the glade file. This is not a good way to
do things, and in the long run, when we move to phase out gtk2 builds,
we should replace this approach with a regular glade file.