Changed ImageUtilities.ResizeImage() function
Previously, this created an image of the desired target height and
width, but most of the "image" area was transparent, and the actual
content of the area only filled as much of the total image area as the
aspect ratio allowed. E.g. you would typically end up with blank bars
above and below the image.

The way that the function works now is that the resultant image's
dimensions will match exactly the target width or height scaled down
according to the original image's aspect ratio.