Removed confusing extension method on GtkGrid
This extension method has the same signature as another function on the
Grid class. It was originally intended for compatibility purposes with
the gtk2 API but it's not actually being used right now and I'm going to
remove it as I think it has the potential to cause problems (we
definitely don't want to call this function by mistake!).