Add CreateDocumentation and Updater back into solution file
So that they are built automatically. I've also had to update the
CreateDocumentation project to target .NET framework 4.7.2; it won't
build against older versions since it has ApsimNG as a dependency which
now targets 4.7.2 as well.

todo: convert updater and autodocs to .net core. Updater currently runs
on winforms, so it needs to be reworked anyway. The only block on
autodocs which I can think of is that ApsimNG isn't totally stable yet
on netcore builds.