Fixed minor bug in markdownview
Previously the GetTags() function always retrieved the TextTags from the
primary TextView. This was fine most of the time, except when measuring
the contents of a cell of a table, in which case we construct a
temporary textview. When this happens, we need to be sure that the tags
we use when inserting text into this textview actually come from that
textview's tags table, otherwise we get a gtk warning message.