Fixed failure in report when run under invariant culture
The date parsing in report is problematic. We kind of expect people to
supply dates in dd-MMM format (e.g. 25-Dec for 25 December).
Unfortunately, calling `DateTime.Parse("25-Dec",
CultureInfo.InvariantCulture)` yields 1 December 2025.

The workaround which I've implemented here is to manually attempt to
parse the date as dd-MMM first, before falling back to the old method.

This was previously causing the unit tests to fail when run under the
invariant culture. (Specifically, the TestAggregationOverYearBoundary
test in ReportTests).