Fixed a gtk warning which was occurring after searching in the treeview
Normally we just need to remove the accelerators when the treeview loses
focus, and re-add them when it regains focus. However, it's possible for
the treeview to gain focus multiple times consecutively without losing
focus in-between, which leads to gtk warnings. Typically this occurs
after using the search functionality. The solution is to keep track of
whether the accelerators are already attached to the window, so that we
only add them when necessary.