Rebuilt PivotTable into ReportPivot
A new pivot table model has been built from the ground up , so that it is more robust. It uses existing views and should work nicely with .NET Core. It should also work with reports outside the context of CLEM, if necessary.

The old PivotTable model has been marked as deprecated, but left in place so as not to break old simulations.

Resolves #6233