Slight refactor of the file-saving code in ExplorerPresenter
There's now only one call to Simulations.Write() - that's from inside
the WriteSimulation() function, which also takes care of saving the
position in the command history.

Previously, we were calling Simulations.Write() without calling
CommandHistory.Save(), which resulted in the CommandHistory thinking
that there were pending changes, causing the user to be prompted to save
the file when closing the tab immediately after having saved the file
manually.