Change autodocs to insert 2 newlines between array values
This way each array value will be rendered as its own paragraph; in the
commonmark world, a single linebreak without 2 or more preceding spaces
is rendered as a 'soft' linebreak - e.g.

foo
bar

becomes

<p>foo
bar</p>

Which will be rendered as a space character by most HTML renderers -
which is not what we want.