Fixed converter code which adds initial water to soil.
The converter has code to add an intial water node to a soil if it
doesn't already have an intial water node. However, initial water can
sometimes be specified from a sample - in such cases we shouldn't really
add an initial water node. The bug could be triggered by copying and
pasting a soil in the gui. This patch kind of allows for this scenario,
but will only help if the sample is called "Initial Water"
(case-insensitively). I've also added a unit test.

There are a number of problems with this entire approach but for now I'm
just going to fix the immediate bug.