Disable apex in second leafcohort in canopyphotosynthesis.apsimx
LeafCohort has a scoped link to an IApex. The infrastructure will now
throw an exception if there are multiple candidates to a link. In this
case, there are two leaf cohorts, and they can both see each other's
child apex, so the link to apex is ambiguous. In fact, previously, the
second leaf cohort would have been linking to the first leaf cohort's
apex, which is why we're making this change.

To get this working properly, we would need to change the source code,
change the link to a child link.