Fixed two scripts in eucalytpus validation
Both scripts were in the wagga simulation and were causing problems
because they had links to ambiguous models. The TreeTransplantRule had a
link to plant, but there were two plants in the zones. Luckily this
happened to be resolving to euc, but it was dependent on the order in
the tree. I've made this a link by name to make it explicit.

The other problem was a scoped link to plant in the auto fertigate
script (again in wagga). Incidentally, the plant variable was called
chickpea. This scoped link was also ambiguous. Fortunately, it wasn't
being used so I removed it.