Nutrient model init code should occur in StartOfSimulation
Previously the initialisation code was occurring in DoSoilOrganicMatter,
which meant that attempts to access variables such as
Nutrient.DenitrifiedN on the first day of the simulation before the
DoSoilOrganicMatter event fired would result in a null reference
exception. This was occurring in the ExtraVariables script in the slurp
test set.