Fixed bug in manager code for setting params in script
If the property value starts with . or [ (ie if it looks like an apsim
path), then we assign to the property whatever exists at the path.
However, this should really only occur if the property type is an IModel
of some sort. Unfortunately, some types like IPlant (which is commonly
used in this context) do not implement IModel, so we also check if the
type is an interface. The main problem we're trying to fix here is not
fetching the value at the apsim path if the property is a string.