SimplePropertyPresenter: save changes when detached
When the Detach() method of SimplePropertyPresenter is called, it needs
to save any outstanding changes. These changse will be from any entry
widgets which have keyboard focus (such changes are applied when the
widget loses focus).

This was causing changes to not be saved when editing a text input in
the manager UI, and then immediately clicking on another model in the
simulations tree.