Fixed a bug where regression line was longer than it should be
This was occurring when the largest x value corresponded to a y value of
NaN (or vice versa). In this case, that particular dot would not be
rendered on the graph but the regression line would nonetheless extend
to this point. The solution is to check for NaNs in the corresponding x
or y value when searching for the max/min values in the series.