Let's try and be more forgiving of empty strings in excel spreadsheets
Previously, a numeric column containing an empty string would be
interpreted as a text column. This then caused the data to be parsed
incorrectly by the graphing package leading to x/y values being
mismatched (ie values from different rows being considered a pair).

The problem here is really with the graphing code and it should be fixed
but this will be more difficult, and the actual problem is actually
caused by the empty string