Fixed keyboard shortcuts in gtk3 editor view
This is the editor component used for manager scripts, reports, etc.
Previously, the keyboard shortcuts (find/replace/compile) weren't
working until the user right-clicked for the first time. This solution
is something of a hack, although I can't think of a better solution
which doesn't require rewriting our menu code here to use GMenu and
GAction.