Added a Prepare() function to Simulation and IRunnable
The idea behind this is that we want the apsim server to be able to hold
a simulations object in memory and run it on-demand without reconnecting
links, events, etc. All of this pre-simulation prep now occurs in the
Prepare() function of Simulation, which is called from the JobRunner via
the IRunnable interface. Possibly this function doesn't actually belong
in this interface, as no other IRunnable implementation implements this
function, but we can think more about that later.

I've also added a ServerJobRunner, which extends the main JobRunner
class, and is designed to be run multiple times on the same simulation
objects.

Finally, I've fixed a number of minor bugs which were occurring when
running a simulation multiple times without cloning it each time.