Added a SubscribeToEvents event
This event is intended to be used by models which dynamically connect to
events. This event is fired once, before the start of the simulation.
The advantage of doing this is so that the apsim server (which runs the
simulations multiple times without cloning them between) doesn't end up
having models connect to events more than once.