Fixed a bug in DynamicContractResolver
This class will determine which properties are serialized when we
call ReflectionUtilities.JsonSerialise(). Normally want to serialize all
properties (including base class properties), so we use
BindingFlags.FlattenHierarchy. The problem with this occurs when the
object to be serialized overrides some properties from its base class -
the list of properties returned by Type.GetProperties() will have
multiple properties with the same name, which is illegal in JSON - a
JSON object cannot have multiple keys (properties) with the same name.

The fix for this is to remove all properties from the list which are
duplicated - keeping the most derived property in each case.