Fixed GrazeForageCrop script in FodderBeet example file
This script modifies one of its inputs (CurrentLocation). The
CurrentLocation variable actually stores the initial location, so I've
renamed it to reflect this and created a new property called
CurrentLocation, to which the initial location is assigned at the start
of the simulation.

Normally, this wouldn't have been a problem, but in an apsim server
scenario, we need to be able to run the simulation multiple times
without cloning it in-between.