Fixed another minor bug with RedClover example defolation script
It was previously modifying one of its inputs
(DefoliationRotationStartDate). This is normally fine, but when run from
an apsim server causes problems when the simulation is rerun without
being cloned.

In this script, I don't think we actually need to modify this variable,
as the greater than check will cover the scenario where the simulation
starts after the defoliation start date.