Fixed generation of .apsimx files for an experiment
...which modifies manager script parameters.

When we generate the list of simulations to be serialized, the changed
manager parameters are applied directly to the script object, but this
object isn't included in the json serialization. Instead, the manager
uses reflection to examine the script and saves its properties to a dict
property of the manager object.

When we run a simulation, the parameters dictionary is not updated, and
we don't particularly care what's in it once the script object is
initialised correctly. In this instance however, we DO care about the
contents of this dict because we're about to serialize the model.