Fixed a bug in JobManager
More of an oversight, really; the job manager objects weren't really
designed to be reused. Now they need to be reused.

Basically this solution is a hack to allow the server job runner to
reuse a job manager (typically a SimulationGroup). Ideally, the
post-simulation stuff should be refactored out of SimulationGroup. What
should really happen is that the numJobsToRun variable should be set to
NumJobs when a run (of all jobs owned by the job manager) starts.