Fixed a bug in Debian/MacOS installer scripts
These installers are currently built on windows. Therefore we need to
remove any windows-specific files from the bin directory before copying
its contents into the install path. The 'windows-specific' files used to
be anything in the DeploymentSupport tree, but it's now changed to
anything in the DeploymentSupport/Windows tree. (Anything under
DeploymentSupport/global needs to stay where it is.)