Tweaked solute's initialisation code
Previously, we called `initial.FindByPath(Name)`, which is a rather
clumsy method of finding a property of the sample. For example, if
someone were to add an irrigation module and name it "Urea", suddenly,
the solute would try and read the irrigation module to get its initial
amount of solute.

I've changed this so we just check for a property on the Sample class.