Change SimulationResetTests to run each sim in parallel
This test will run most of the example files shipped with apsim to
verify that they work with the server feature. This is obviously rather
slow, but the simulations all run in parallel with each other (as
separate test cases) and in parallel with all other tests in this
assembly.