Changed .apsimx file reading API
Now instead of an out parameter containing exceptions encountered, the
caller must specify an action to take when an exception is thrown. This
allows the UI to attempt to continue reading the file, and other callers
to abort early upon failure.

I've also added a boolean parameter which causes the calls to the read
models' OnCreated() methods to occur in a background thread. These calls
take far longer than actually reading/deserializing the file, so in a UI
context it is not desirable to block until they finish running. In other
contexts, we can ensure that the model has been initialised correctly
upon returning by passing in a false value for this parameter.