Fixed a race condition in the data store writer
The WaitForIdle() function will wait until there are no jobs
and idle is set to true. Therefore, we should update the value
of idle *before* removing this command from the commands list.
Otherwise, we could end up in a situation where idle is (briefly)
set to true (because a command was recently added), and the commands
list is empty, which would cause WaitForIdle() to return, even
though the job runner actually hasn't finished running this command.