Tweak salb in orchard prototype soils to avoid triggingering NSE test
soil_albedo has been removed from microclimate - microclimate will
instead use salb in the soil water model. This has caused the orchard
prototype to fail its NSE test because the soil_albedo in microclimate
in all sims was set to 0.3, whereas it's now using the value of 0.18 in
the soil. I've changed the values in the soils to 0.3 to match the old
behaviour in micrcoclimate, although the soil temp model will likely
have different results now.