Add the ability for DataStoreWriter to run a clean command internally
This needs to be called from SimulationGroup before running simulations.
However, we don't want to add this to the simulation group's list of
jobs, otherwise the status (as displayed in the GUI) will include the
clean command in the total number of jobs to be run, which is implied to
be only the simulations.