Ensure that post-simulation tools clean existing data
...because DataStoreWriter.Clean() is called if and only if simulations
are going to be run. So if the user does a "refresh" operation in the
GUI (ie to rerun post-simulation tools), then no data cleaning will
occur, and so the post-simulation tools are really responsible for
cleaning up after themselves anyway.

To this end, I've changed the second argument to
`DataStoreWriter.WriteTable()` (which is is a boolean flag to delete
existing data); I've set the default value to true, because the only
times where we *don't* want to delete existing data is in the few
situations where we're calling from a running simulation (ie report).
Deleting existing data is the default and safest option, and should only
be disabled in well-tested cases as an optimisation (as in the case of
Report).