Restructed DataStoreView.glade to avoid warnings in gtk2 builds
Certain properties such as halign and margin-bottom were added to the
widget class in gtk3; including these properties in the .glade file
causes warnings to occur in gtk2 builds. Fortunately, there are ways to
achieve the same result which work in both gtk2 and 3 builds;

- Xalign property on label class intead of halign
- Using a GtkAlignment object instead of setting margin-bottom

These changes could and probably should be reverted when we drop gtk2
support.