Fixed a bug in GraphPage code
When we generate series definitions from IGraphable instances which are
children of a series, they should only be given the series definitions
from their parent series.

For example, if we have an EventNamesOnGraph instance which is a child
of a Series, when we ask the EventNamesOnGraph to generate its series
definitions, it should only be passed in data from its parent series,
rather than from all series in the graph (which is what we were doing
previously). The result was that the results from the child IGraphable
could depend upon the order of the series in the simulations tree.