Fixed a bug in Series code
We were passing in a cloned list, which was supposed to be both an input
and an output. I've changed this function so that it takes an
IEnumerable<SeriesDefinition> as input, and returns an
IEnumerable<SeriesDefinition>. This way the callee doesn't attempt to
(or need to) modify the list which is passed in - which is really quite
a sin.