Use OxyPlot.SkiaSharp for graph rendering in autodocs
The builtin svg exporter is quite primitive, and the resultant images
don't look great.

Note that using the OxyPlot.SkiaSharp package has required me to bump
the version of oxyplot which we are using, which has introduced a large
number of breaking changes, hence the changes to ApsimNG in this commit.
In the mid-term, we will want to rewrite the GraphView anyway, in order
to take advantage of the graph->oxyplot pipeline in APSIM.Interop.