Ensure that large images don't go into existing paragraphs
Theoretically, we could resize them a bit, but we don't really have a
way of knowing how much they should be resized such that they fit well
with the surrounding content.

I've opted for the simple approach and given images their own paragraph
(their own page, basically) if the image is almost as tall as the page.