Add soybean phenology diagram as resource file
This used to be referenced from a memo in Soybean.apsimx, but that memo
has been moved into Soybean.json. As a result, relying on a relative
path no longer works. This would work if we moved the image into the
Models/Resources directory, but I'd rather reserve that location for
resource models. Embedding the image into the APSIM.Interop assembly
works fine. (Putting it in the Models assembly would also be fine.)