Fixed regression lines on graph
In many cases, x or y (but not both) may be NaN. This is common in P/O
tables. In such cases, we cannot add the y-value into the list of
y-values but ignore the x-value for being NaN. It appears to be safe to
accept NaNs here (they will be ignored by the stats code) so that's
what I'm going to do.