Temp hack to fix croptimizr when running locally
When running inside docker, the R script needs to be built with relative
paths (there are multiple problems with absolute paths). This works
reasonably well by setting the working directory in the container. In an
ideal world we could use the same approach when running the R code
outside of docker as well. Unfortunately, the R wrapper assumes control
over the working dir and relies on this in order to force the use of a
custom .Rprofile file. While it should be possible to work around this,
I think that such a change would be beyond the scope of this PR.