Allow sending many strings, and dates.
Strings when sent in a READ command were not delimited in any way, so the client couldn't break them up again.
They are now prefixed with an integer containing their length.

This also implements a protocol for sending dates -- see the assumption comment near the definition of 'dateFormat'.
The dates are converted to strings in ISO format (e.g. 2022-03-04) before sending.