Fixed a crash which sometimes happened when clicking away from report
This was happening because the intellisense view's main widget is a
Window, which is a top-level widget. When we call Dispose() on a
top-level widget, GtkSharp will automatically call Destroy() for us.
Calling Destroy() twice is not a good idea, so we should only manually
call Destroy() if the widget is not a top-level window.

Prior to this change, I was consistently getting either a crash or the
following message in stdout when clicking away from a report node:

GLib-GObject-CRITICAL **: 12:57:15.718: g_object_remove_toggle_ref:
assertion 'G_IS_OBJECT (object)' failed