Fixed reporting in apsim server workloads
SubscribeToEvents is only called once by the server - and that's
immediately upon application launch. If we disconnect events at
EndOfSimulation, they'll never be reconnected.

In a GUI run, SubscribeToEvents is called once per run, so the events
would get reconnected, but we're running a cloned copy of the
simulations which is discarded after the run, so it really doesn't
matter if we don't disconnect the events.

In a CLI run, we only run once and then exit, so it doesn't really
matter whether we detach them, so long as we connect them once in the
first place.