Corrected position tracking in stream reader (again)
When updating the offset within the file, we use

new offset = offset within file before read + # bytes read

The offset within the file before the read is the offset of the buffer
within the file (in bytes) + the offset within the buffer (in bytes).
The offset of the buffer within the file takes into account the
byte-order mark at the start of the stream (if any is present). I've
added another global variable to track the position within the buffer
(in bytes), which is accumulated as we read through the file, and should
be modified whenever position is modified.

I've also added unit tests to ensure that this is all working as
expected.