Fixed culture/locale settings in water presenter
In general, we need to read/write to/from the GUI using the current culture, and
read/write to .apsimx files using the invariant culture. This allows users to
view data "natively" and allow them to run files generated in a different locale
(e.g. sent to them by a friend).