Fixed culture settings used in WaterPresenter
Previously, we were attempting to parse double values (coming from the
GUI) using the invariant culture. This is a sure recipe for disaster;
we need to read/write to the GUI using the *current* culture, and
read/write to/from the .apsimx file and .db file using the *invariant*
culture. This ensures that anyone can open an .apsimx file and run it
and achieve basically the same results, no matter their system locale.