Fixed numpad in SheetView
Typing numbers via the numpad should cause the SheetView to enter edit
mode, when editing is enabled. This code was taken from the old GridView
widget, which, while not the most stable thing in the world, is at least
fairly well-tested. (GridView.cs:1466)

I've tested this locally and it seems to solve the problem.