Prevent Gtk-CRITICAL messages from being generated
...when hovering over empty space in a ListView widget. When we try to
do this, calling gtk_tree_view_get_path_at_pos() with the current
coordinates will return a NULL GtkTreePath pointer (as it should - what
valid path could refer to this location?). If we then pass a null
pointer into gtk_tree_model_get_iter(), we are abusing the API, which,
in a best case scenario, will result in a warning message from being
written to stdout.

The easy solution here is to do nothing.